moved mock backend to own file; fix translation variables only replacing first occurrence
